+++ shared/Patch.h
@@ -6,53 +6,56 @@
 */
 
 #pragma once
+#include <cstdint>
+
 #include "../injector/assembly.hpp"
 #include "../injector/injector.hpp"
 #include "DynAddress.h"
 #include <vector>
+#include <memory>
 
 namespace plugin {
 
 class patch {
 public:
-    static void Nop(int address, size_t size, bool vp = true);
-    static void RedirectCall(int address, void *func, bool vp = true);
-    static void RedirectJump(int address, void *func, bool vp = true);
-    static void SetChar(int address, char value, bool vp = true);
-    static void SetUChar(int address, unsigned char value, bool vp = true);
-    static void SetShort(int address, short value, bool vp = true);
-    static void SetUShort(int address, unsigned short value, bool vp = true);
-    static void SetInt(int address, int value, bool vp = true);
-    static void SetUInt(int address, unsigned int value, bool vp = true);
-    static void SetFloat(int address, float value, bool vp = true);
-    static void SetPointer(int address, void *value, bool vp = true);
-    static char GetChar(int address, bool vp = true);
-    static unsigned char GetUChar(int address, bool vp = true);
-    static short GetShort(int address, bool vp = true);
-    static unsigned short GetUShort(int address, bool vp = true);
-    static int GetInt(int address, bool vp = true);
-    static unsigned int GetUInt(int address, bool vp = true);
-    static float GetFloat(int address, bool vp = true);
-    static void *GetPointer(int address, bool vp = true);
-    static void SetRaw(int address, void* value, size_t size, bool vp = true);
-    static void GetRaw(int address, void* ret, size_t size, bool vp = true);
-    static void RedirectShortJump(int address, void* dest = nullptr, bool vp = true);
-    static void PutRetn(int address, unsigned short BytesToPop = 0, bool vp = true);
-    static void PutRetn0(int address, unsigned short BytesToPop = 0, bool vp = true);
-    static void PutRetn1(int address, unsigned short BytesToPop = 0, bool vp = true);
+    static void Nop(uintptr_t address, size_t size, bool vp = true);
+    static void RedirectCall(uintptr_t address, void *func, bool vp = true);
+    static void RedirectJump(uintptr_t address, void *func, bool vp = true);
+    static void SetChar(uintptr_t address, char value, bool vp = true);
+    static void SetUChar(uintptr_t address, unsigned char value, bool vp = true);
+    static void SetShort(uintptr_t address, short value, bool vp = true);
+    static void SetUShort(uintptr_t address, unsigned short value, bool vp = true);
+    static void SetInt(uintptr_t address, uintptr_t value, bool vp = true);
+    static void SetUInt(uintptr_t address, uintptr_t value, bool vp = true);
+    static void SetFloat(uintptr_t address, float value, bool vp = true);
+    static void SetPointer(uintptr_t address, void *value, bool vp = true);
+    static char GetChar(uintptr_t address, bool vp = true);
+    static unsigned char GetUChar(uintptr_t address, bool vp = true);
+    static short GetShort(uintptr_t address, bool vp = true);
+    static unsigned short GetUShort(uintptr_t address, bool vp = true);
+    static uintptr_t GetInt(uintptr_t address, bool vp = true);
+    static uintptr_t GetUInt(uintptr_t address, bool vp = true);
+    static float GetFloat(uintptr_t address, bool vp = true);
+    static void *GetPointer(uintptr_t address, bool vp = true);
+    static void SetRaw(uintptr_t address, void* value, size_t size, bool vp = true);
+    static void GetRaw(uintptr_t address, void* ret, size_t size, bool vp = true);
+    static void RedirectShortJump(uintptr_t address, void* dest = nullptr, bool vp = true);
+    static void PutRetn(uintptr_t address, unsigned short BytesToPop = 0, bool vp = true);
+    static void PutRetn0(uintptr_t address, unsigned short BytesToPop = 0, bool vp = true);
+    static void PutRetn1(uintptr_t address, unsigned short BytesToPop = 0, bool vp = true);
 
     template <typename T>
-    static void Set(int address, T value, bool vp = true) {
+    static void Set(uintptr_t address, T value, bool vp = true) {
         injector::WriteMemory(GetGlobalAddress(address), value, vp);
     }
 
     template <typename T>
-    static T Get(int address, bool vp = true) {
+    static T Get(uintptr_t address, bool vp = true) {
         return injector::ReadMemory<T>(GetGlobalAddress(address), vp);
     }
 
     template <typename T>
-    static T TranslateCallOffset(int address) {
+    static T TranslateCallOffset(uintptr_t address) {
         if (GetUChar(address) == 0xE8) {
             auto offset = GetInt(address + 1);
             offset += (GetGlobalAddress(address) + 5);
@@ -62,7 +65,7 @@ public:
     }
 
     template <typename T>
-    static T TranslateJumpOffset(int address) {
+    static T TranslateJumpOffset(uintptr_t address) {
         if (GetUChar(address) == 0xE9) {
             auto offset = GetInt(address + 1);
             offset += (GetGlobalAddress(address) + 5);
@@ -71,47 +74,47 @@ public:
         return nullptr;
     }
 
-    static void ReplaceFunction(int address, void *func, bool vp = true);
-    static void ReplaceFunctionCall(int address, void *func, bool vp = true);
+    static void ReplaceFunction(uintptr_t address, void *func, bool vp = true);
+    static void ReplaceFunctionCall(uintptr_t address, void *func, bool vp = true);
 
-    static void SetPointer(std::vector<int> const& addresses, void* value, bool vp = true) {
+    static void SetPointer(std::vector<uintptr_t> const& addresses, void* value, bool vp = true) {
         for (auto& address : addresses) {
             SetPointer(address, value, vp);
         }
     }
 
     template <typename T>
-    static void Set(std::vector<int> const& addresses, T value, bool vp = true) {
+    static void Set(std::vector<uintptr_t> const& addresses, T value, bool vp = true) {
         for (auto& address : addresses) {
             Set(address, value, vp);
         }
     }
 
-    static void SetInt(std::vector<int> const& addresses, int value, bool vp = true) {
+    static void SetInt(std::vector<uintptr_t> const& addresses, uintptr_t value, bool vp = true) {
         for (auto& address : addresses) {
             SetInt(address, value, vp);
         }
     }
 
-    static void SetFloat(std::vector<int> const& addresses, float value, bool vp = true) {
+    static void SetFloat(std::vector<uintptr_t> const& addresses, float value, bool vp = true) {
         for (auto& address : addresses) {
             SetFloat(address, value, vp);
         }
     }
 
-    static void RedirectJump(std::vector<int> const& addresses, void* func, bool vp = true) {
+    static void RedirectJump(std::vector<uintptr_t> const& addresses, void* func, bool vp = true) {
         for (auto& address : addresses) {
             RedirectJump(address, func, vp);
         }
     }
 
-    static void RedirectCall(std::vector<int> const& addresses, void* func, bool vp = true) {
+    static void RedirectCall(std::vector<uintptr_t> const& addresses, void* func, bool vp = true) {
         for (auto& address : addresses) {
             RedirectCall(address, func, vp);
         }
     }
 
-    static void Nop(std::vector<int> const& addresses, size_t size, bool vp = true) {
+    static void Nop(std::vector<uintptr_t> const& addresses, size_t size, bool vp = true) {
         for (auto& address : addresses) {
             Nop(address, size, vp);
         }
@@ -122,7 +125,7 @@ public:
     static inline std::unordered_map<uint32_t, void(*)(RegPack&)> staticHookMap;
     static inline int32_t staticHookIndices = 0;
 
-    static void StaticHook(uint32_t start, uint32_t end, void (*func)(RegPack&)) {
+    static void StaticHook(uintptr_t start, uintptr_t end, void (*func)(RegPack&)) {
         staticHookMap[staticHookIndices++] = func;
         struct staticHook {
             int32_t id = staticHookIndices - 1;
@@ -133,7 +136,7 @@ public:
     }
 
     template <typename T>
-    static void StaticHook(uint32_t start, uint32_t end) {
+    static void StaticHook(uintptr_t start, uintptr_t end) {
         injector::MakeInline<T>(start, end);
     }
 };
